cleanup and using featurecollection where I can
